Refactoring
GPGraph.hpp separated into GraphNode.hpp, Graph.hpp, GraphBuilder.hpp
