--- source/Group7GP/GPGraph.hpp
@@ -1,340 +1,0 @@
-
-#pragma once
-
-#include <algorithm>
-#include <iostream>
-#include <memory>
-#include <numeric>
-#include <string>
-#include <unordered_map>
-#include <vector>
-
-#include "../core/AgentBase.hpp"
-
-namespace cowboys {
-    using NodeFunction = double (*)(const std::vector<double> &);
-
-    /// @brief Sum all inputs.
-    /// @param inputs Vector of doubles representing the inputs.
-    /// @return The sum of all inputs.
-    double Sum(const std::vector<double> &inputs) { return std::accumulate(inputs.begin(), inputs.end(), 0.0); }
-
-    /// @brief Check if all inputs are not equal to 1.
-    /// @param inputs Vector of doubles representing the inputs.
-    /// @return 1 if all inputs are not equal to 0, 0 otherwise.
-    double And(const std::vector<double> &inputs) {
-        for (auto &input : inputs) {
-            if (input == 0.)
-                return 0.;
-        }
-        return 1.;
-    }
-
-    /// @brief Check if any of the inputs besides the first are equal to the
-    /// first input.
-    /// @param inputs Vector of doubles representing the inputs.
-    /// @return 1 if any of the inputs besides the first are equal to the first
-    /// input, 0 otherwise.
-    double AnyEq(const std::vector<double> &inputs) {
-        for (size_t i = 1; i < inputs.size(); ++i) {
-            if (inputs.at(0) == inputs.at(i))
-                return 1.;
-        }
-        return 0.;
-    }
-
-    /// @brief Check if the first input is equal to 0.
-    /// @param inputs Vector of doubles representing the inputs.
-    /// @return 1 if the first input is equal to 0 or there are no inputs, 0
-    /// otherwise.
-    double Not(const std::vector<double> &inputs) { return (inputs.size() == 0) || (inputs.at(0) == 0.) ? 1. : 0.; }
-
-    /// @brief Returns the input with index 0 if the condition (input with index
-    /// 1) is not 0.
-    /// @param inputs Vector of doubles representing the inputs.
-    /// @return The value of the input with index 0 if the condition (input with
-    /// index 1) is not 0, 0 otherwise.
-    double Gate(const std::vector<double> &inputs) { return inputs.at(1) != 0. ? inputs.at(0) : 0.; }
-
-    class GraphNode {
-      protected:
-        /// The input nodes to this node.
-        std::vector<std::shared_ptr<GraphNode>> inputs;
-
-        /// The function that operates on the outputs from a node's input nodes.
-        NodeFunction function_pointer{nullptr};
-
-        /// The default output of this node.
-        double output{0};
-
-      public:
-        GraphNode() = default;
-        virtual ~GraphNode() = default;
-
-        /// TODO: Check guidelines for this
-        GraphNode(double default_value) : output{default_value} {}
-        GraphNode(NodeFunction function) : function_pointer{function} {}
-
-        std::string name;
-
-        /// TODO: Cache outputs
-        virtual double GetOutput() const {
-            // Invoke function pointer if it exists
-            if (function_pointer != nullptr) {
-                std::vector<double> input_values;
-                for (auto &input : inputs) {
-                    input_values.push_back(input->GetOutput());
-                }
-
-                try {
-                    return function_pointer(input_values);
-                } catch (const std::out_of_range &e) {
-                    return 0.0;
-                }
-            }
-
-            // Default output
-            return output;
-        }
-
-        void SetFunctionPointer(NodeFunction function) { function_pointer = function; }
-
-        std::shared_ptr<GraphNode> GetInput(size_t input_idx) const {
-            if (inputs.size() == 0) {
-                // Throw error if no inputs
-                throw std::runtime_error("No inputs");
-            }
-            if (input_idx >= inputs.size())
-                // Return last input if out of bounds
-                return inputs.back();
-            else
-                return inputs.at(input_idx);
-        }
-
-        void AddInput(std::shared_ptr<GraphNode> node) { inputs.push_back(node); }
-        void AddInputs(const std::vector<std::shared_ptr<GraphNode>> &nodes) {
-            inputs.insert(inputs.end(), nodes.begin(), nodes.end());
-        }
-        void SetInputs(const std::vector<std::shared_ptr<GraphNode>> &nodes) { inputs = nodes; }
-        void SetOutput(double value) { output = value; }
-    };
-
-    using GraphLayer = std::vector<std::shared_ptr<GraphNode>>;
-
-    class Graph {
-      protected:
-        /// Actions that can be taken.
-        std::vector<size_t> actions;
-
-        /// Layers of nodes in the graph.
-        std::vector<GraphLayer> layers;
-
-      public:
-        Graph(const std::vector<size_t> &action_vec) : actions{action_vec} { assert(actions.size() > 0); }
-        ~Graph() = default;
-
-        size_t GetNodeCount() const {
-            size_t count = 0;
-            for (const auto &layer : layers) {
-                count += layer.size();
-            }
-            return count;
-        }
-
-        size_t GetLayerCount() const { return layers.size(); }
-
-        size_t MakeDecision(const std::vector<double> &inputs) {
-            if (layers.size() == 0)
-                return actions.at(0);
-
-            // Set inputs of input layer
-            size_t i = 0;
-            for (auto &node : layers[0]) {
-                double input = 0;
-                if (i < inputs.size())
-                    input = inputs.at(i);
-                node->SetOutput(input);
-                ++i;
-            }
-
-            // Get output of last layer
-            std::vector<double> outputs;
-            for (auto &node : layers.back()) {
-                outputs.push_back(node->GetOutput());
-            }
-
-            std::cout << "Outputs: ";
-            for (auto &output : outputs) {
-                std::cout << output << " ";
-            }
-            std::cout << std::endl;
-
-            // Choose the action with the highest output
-            auto max_output = std::max_element(outputs.begin(), outputs.end());
-            size_t index = std::distance(outputs.begin(), max_output);
-
-            // If index is out of bounds, return the last action
-            size_t action = 0;
-            if (index >= actions.size())
-                action = actions.back();
-            else
-                action = actions.at(index);
-            std::cout << "Action: " << action << std::endl;
-            return action;
-        }
-
-        void AddLayer(const GraphLayer &layer) { layers.push_back(layer); }
-    };
-
-    /// @brief Encodes the actions from an agent's action map into a vector of
-    /// size_t, representing action IDs.
-    /// @param action_map The action map from the agent.
-    /// @return A vector of size_t, representing action IDs.
-    std::vector<size_t> EncodeActions(const std::unordered_map<std::string, size_t> &action_map) {
-        std::vector<size_t> actions;
-        for (const auto &[action_name, action_id] : action_map) {
-            actions.push_back(action_id);
-        }
-        // Sort the actions so that they are in a consistent order.
-        std::sort(actions.begin(), actions.end());
-        return actions;
-    }
-
-    /// @brief Translates state into nodes for the decision graph.
-    /// @return A vector of doubles for the decision graph.
-    std::vector<double> EncodeState(const cse491::WorldGrid &grid, const cse491::type_options_t & /*type_options*/,
-                                    const cse491::item_set_t & /*item_set*/, const cse491::agent_set_t & /*agent_set*/,
-                                    const cse491::Entity *agent,
-                                    const std::unordered_map<std::string, double> &extra_agent_state) {
-        /// TODO: Implement this function properly.
-        std::vector<double> inputs;
-
-        auto current_position = agent->GetPosition();
-
-        double current_state = grid.At(current_position);
-        double above_state = grid.IsValid(current_position.Above()) ? grid.At(current_position.Above()) : 0.;
-        double below_state = grid.IsValid(current_position.Below()) ? grid.At(current_position.Below()) : 0.;
-        double left_state = grid.IsValid(current_position.ToLeft()) ? grid.At(current_position.ToLeft()) : 0.;
-        double right_state = grid.IsValid(current_position.ToRight()) ? grid.At(current_position.ToRight()) : 0.;
-
-        double prev_action = extra_agent_state.at("previous_action");
-
-        inputs.insert(inputs.end(), {prev_action, current_state, above_state, below_state, left_state, right_state});
-
-        return inputs;
-    }
-
-    class GraphBuilder {
-      protected:
-        /// Action ids from the agent's action map
-        std::vector<size_t> actions;
-
-      public:
-        GraphBuilder(const std::unordered_map<std::string, size_t> &action_map) : actions{EncodeActions(action_map)} {}
-        ~GraphBuilder() = default;
-
-        std::unique_ptr<Graph> CartesianGraph(size_t num_inputs, size_t num_outputs, size_t num_layers,
-                                              size_t num_nodes_per_layer) {
-            auto decision_graph = std::make_unique<Graph>(actions);
-
-            // Input layer
-            GraphLayer input_layer;
-            for (size_t i = 0; i < num_inputs; ++i) {
-                input_layer.push_back(std::make_shared<GraphNode>(0));
-            }
-            decision_graph->AddLayer(input_layer);
-
-            // Middle Layers
-            for (size_t i = 0; i < num_layers; ++i) {
-                GraphLayer layer;
-                for (size_t j = 0; j < num_nodes_per_layer; ++j) {
-                    layer.push_back(std::make_shared<GraphNode>(0));
-                }
-                decision_graph->AddLayer(layer);
-            }
-
-            // Action layer
-            GraphLayer output_layer;
-            for (size_t i = 0; i < num_outputs; ++i) {
-                output_layer.push_back(std::make_shared<GraphNode>(0));
-            }
-            decision_graph->AddLayer(output_layer);
-
-            return decision_graph;
-        }
-
-        /// @brief Creates a decision graph for pacing up and down in a
-        /// MazeWorld. Assumes that the inputs are in the format: prev_action,
-        /// current_state, above_state, below_state, left_state, right_state
-        /// @param action_vec Assumes that the action outputs are in the format:
-        /// up, down, left, right
-        /// @return
-        std::unique_ptr<Graph> VerticalPacer() {
-            auto decision_graph = std::make_unique<Graph>(actions);
-
-            GraphLayer input_layer;
-            std::shared_ptr<GraphNode> prev_action = std::make_shared<GraphNode>(0);
-            std::shared_ptr<GraphNode> current_state = std::make_shared<GraphNode>(0);
-            std::shared_ptr<GraphNode> above_state = std::make_shared<GraphNode>(0);
-            std::shared_ptr<GraphNode> below_state = std::make_shared<GraphNode>(0);
-            std::shared_ptr<GraphNode> left_state = std::make_shared<GraphNode>(0);
-            std::shared_ptr<GraphNode> right_state = std::make_shared<GraphNode>(0);
-            input_layer.insert(input_layer.end(),
-                               {prev_action, current_state, above_state, below_state, left_state, right_state});
-            decision_graph->AddLayer(input_layer);
-
-            // state == 1 => floor which is walkable
-            GraphLayer obstruction_layer;
-            std::shared_ptr<GraphNode> up_not_blocked = std::make_shared<GraphNode>(AnyEq);
-            up_not_blocked->AddInputs(GraphLayer{above_state, std::make_shared<GraphNode>(1)});
-            std::shared_ptr<GraphNode> down_not_blocked = std::make_shared<GraphNode>(AnyEq);
-            down_not_blocked->AddInputs(GraphLayer{below_state, std::make_shared<GraphNode>(1)});
-            obstruction_layer.insert(obstruction_layer.end(), {up_not_blocked, down_not_blocked});
-            decision_graph->AddLayer(obstruction_layer);
-
-            // Separate previous action into up and down nodes
-            GraphLayer prev_action_layer;
-            std::shared_ptr<GraphNode> up_prev_action = std::make_shared<GraphNode>(AnyEq);
-            up_prev_action->AddInputs(GraphLayer{prev_action, std::make_shared<GraphNode>(1)});
-            std::shared_ptr<GraphNode> down_prev_action = std::make_shared<GraphNode>(AnyEq);
-            down_prev_action->AddInputs(GraphLayer{prev_action, std::make_shared<GraphNode>(2)});
-            prev_action_layer.insert(prev_action_layer.end(), {up_prev_action, down_prev_action});
-            decision_graph->AddLayer(prev_action_layer);
-
-            GraphLayer moving_layer;
-            // If up_not_blocked and up_prev_action ? return 1 : return 0
-            // If down_not_blocked and down_prev_action ? return 1 : return 0
-            std::shared_ptr<GraphNode> keep_up = std::make_shared<GraphNode>(And);
-            keep_up->AddInputs(GraphLayer{up_not_blocked, up_prev_action});
-            std::shared_ptr<GraphNode> keep_down = std::make_shared<GraphNode>(And);
-            keep_down->AddInputs(GraphLayer{down_not_blocked, down_prev_action});
-            moving_layer.insert(moving_layer.end(), {keep_up, keep_down});
-            decision_graph->AddLayer(moving_layer);
-
-            // If down_blocked, turn_up
-            // If up_blocked, turn_down
-            GraphLayer turn_layer;
-            std::shared_ptr<GraphNode> turn_up = std::make_shared<GraphNode>(Not);
-            turn_up->AddInputs(GraphLayer{down_not_blocked});
-            std::shared_ptr<GraphNode> turn_down = std::make_shared<GraphNode>(Not);
-            turn_down->AddInputs(GraphLayer{up_not_blocked});
-            turn_layer.insert(turn_layer.end(), {turn_up, turn_down});
-            decision_graph->AddLayer(turn_layer);
-
-            // Output layer, up, down, left, right
-            GraphLayer action_layer;
-            // move up = keep_up + turn_up,
-            // move down = keep_down + turn_down,
-            std::shared_ptr<GraphNode> up = std::make_shared<GraphNode>(Sum);
-            up->AddInputs(GraphLayer{keep_up, turn_up});
-            std::shared_ptr<GraphNode> down = std::make_shared<GraphNode>(Sum);
-            down->AddInputs(GraphLayer{keep_down, turn_down});
-            std::shared_ptr<GraphNode> left = std::make_shared<GraphNode>(0);
-            std::shared_ptr<GraphNode> right = std::make_shared<GraphNode>(0);
-            action_layer.insert(action_layer.end(), {up, down, left, right});
-            decision_graph->AddLayer(action_layer);
-
-            return decision_graph;
-        }
-    };
-} // namespace cowboys
