+++ tests/unit/gp_agents/GraphNode.cpp
@@ -1,15 +1,16 @@
 /**
  * This file is part of the Fall 2023, CSE 491 course project.
- * @brief Unit tests for Data.hpp in source/core
+ * @brief Unit tests for source/Group7GP/GraphNode.hpp
  **/
 
 // Catch2
 #define CATCH_CONFIG_MAIN
 #include <catch2/catch_all.hpp>
 
-#include "Group7GP/GPGraph.hpp"
+#include "Group7GP/GraphNode.hpp"
 
 using namespace cowboys;
+
 TEST_CASE("GraphNode", "[group7][graphnode]") {
     SECTION("Empty GraphNode") {
         GraphNode node;
@@ -128,56 +129,5 @@ TEST_CASE("GraphNode function set", "[group7][functionset]") {
 
         condition->SetOutput(0);
         CHECK(node.GetOutput() == 0);
-    }
-}
-
-TEST_CASE("Graph", "[group7][graph]") {
-    SECTION("Empty Graph") {
-        std::vector<size_t> actions{1, 2, 3, 4};
-        Graph graph(actions);
-        CHECK(graph.GetLayerCount() == 0);
-        CHECK(graph.GetNodeCount() == 0);
-
-        // Default output the first action
-        std::vector<double> inputs = {1.0, 2.0, 3.0};
-        CHECK(graph.MakeDecision(inputs) == actions.at(0));
-
-        std::vector<size_t> actions2{10, 2, 3, 4};
-        Graph graph2(actions2);
-        CHECK(graph2.MakeDecision(inputs) == actions2.at(0));
-    }
-
-    SECTION("Non-Empty Graph") {
-        Graph graph({1, 2, 3});
-        GraphLayer layer1;
-        layer1.push_back(std::make_shared<GraphNode>());
-        layer1.push_back(std::make_shared<GraphNode>());
-        layer1.push_back(std::make_shared<GraphNode>());
-        graph.AddLayer(layer1);
-        CHECK(graph.GetLayerCount() == 1);
-        CHECK(graph.GetNodeCount() == 3);
-    }
-}
-TEST_CASE("Cartesian Graph", "[group7][graph][cartesian]") {
-    SECTION("Cartesian Graph construction") {
-        constexpr size_t INPUT_SIZE = 10000;
-        constexpr size_t NUM_LAYERS = 10;
-        constexpr size_t NUM_NODES_PER_LAYER = 10;
-        constexpr size_t NUM_NODE_INPUTS = 10;
-        constexpr size_t LAYERS_BACK = 2;
-        auto action_map = std::unordered_map<std::string, size_t>{{"up", 0}, {"down", 1}, {"left", 2}, {"right", 3}};
-        GraphBuilder builder(action_map);
-
-        auto graph = builder.CartesianGraph(INPUT_SIZE, action_map.size(), NUM_LAYERS, NUM_NODES_PER_LAYER);
-
-        // Input layer + middle layers + output layer
-        size_t expected_layer_count = NUM_LAYERS + 2;
-        CHECK(graph->GetLayerCount() == expected_layer_count);
-
-        size_t num_input_nodes = INPUT_SIZE;
-        size_t num_middle_nodes = NUM_LAYERS * NUM_NODES_PER_LAYER;
-        size_t num_output_nodes = action_map.size();
-        size_t expected_node_count = num_input_nodes + num_middle_nodes + num_output_nodes;
-        CHECK(graph->GetNodeCount() == expected_node_count);
     }
 }
