+++ tests/unit/gp_agents/targets.txt
@@ -1,2 +1,4 @@
 agent
-GPGraph
+GraphNode
+Graph
+GraphBuilder
